Move classes from jflex-testsuite-maven-plugin in package jflex.maven.plugin.testsuite;
Follow-up of #458
