+++ testsuite/jflex-testsuite-maven-plugin/src/main/java/jflex/maven/plugin/testsuite/CustomClassLoader.java
@@ -1,4 +1,4 @@
-package jflextest;
+package jflex.maven.plugin.testsuite;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
