+++ testsuite/jflex-testsuite-maven-plugin/src/main/java/jflex/maven/plugin/testsuite/PomUtils.java
@@ -1,4 +1,4 @@
-package jflextest;
+package jflex.maven.plugin.testsuite;
 
 import java.io.File;
 import java.io.FileNotFoundException;
