+++ testsuite/jflex-testsuite-maven-plugin/src/main/jflex/TestLoader.flex
@@ -1,4 +1,4 @@
-package jflextest;
+package jflex.maven.plugin.testsuite;
 
 import java.util.*;
 import jflex.sym;
